+++ src/Oauth1.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace GuzzleHttp\Subscriber\Oauth;
 
 use GuzzleHttp\Psr7\Query;
@@ -9,26 +11,27 @@ use Psr\Http\Message\RequestInterface;
  * OAuth 1.0 signature plugin.
  *
  * Portions of this code comes from HWIOAuthBundle and a Guzzle 3 pull request:
+ *
  * @author Alexander <iam.asm89@gmail.com>
  * @author Joseph Bielawski <stloyd@gmail.com>
  * @author Francisco Facioni <fran6co@gmail.com>
- * @link https://github.com/hwi/HWIOAuthBundle
- * @link https://github.com/guzzle/guzzle/pull/563 Original Guzzle 3 pull req.
  *
- * @link http://oauth.net/core/1.0/#rfc.section.9.1.1 OAuth specification
+ * @see https://github.com/hwi/HWIOAuthBundle
+ * @see https://github.com/guzzle/guzzle/pull/563 Original Guzzle 3 pull req.
+ * @see https://oauth.net/core/1.0/#rfc.section.9.1.1 OAuth specification
  */
 class Oauth1
 {
     /**
-     * Consumer request method constants. See http://oauth.net/core/1.0/#consumer_req_param
+     * Consumer request method constants. See https://oauth.net/core/1.0/#consumer_req_param
      */
-    const REQUEST_METHOD_HEADER = 'header';
-    const REQUEST_METHOD_QUERY  = 'query';
+    public const REQUEST_METHOD_HEADER = 'header';
+    public const REQUEST_METHOD_QUERY = 'query';
 
-    const SIGNATURE_METHOD_HMAC       = 'HMAC-SHA1';
-    const SIGNATURE_METHOD_HMACSHA256 = 'HMAC-SHA256';
-    const SIGNATURE_METHOD_RSA        = 'RSA-SHA1';
-    const SIGNATURE_METHOD_PLAINTEXT  = 'PLAINTEXT';
+    public const SIGNATURE_METHOD_HMAC = 'HMAC-SHA1';
+    public const SIGNATURE_METHOD_HMACSHA256 = 'HMAC-SHA256';
+    public const SIGNATURE_METHOD_RSA = 'RSA-SHA1';
+    public const SIGNATURE_METHOD_PLAINTEXT = 'PLAINTEXT';
 
     /** @var array Configuration settings */
     private $config;
@@ -43,26 +46,29 @@ class Oauth1
      * - callback: OAuth callback
      * - consumer_key: Consumer key string. Defaults to "anonymous".
      * - consumer_secret: Consumer secret. Defaults to "anonymous".
-     * - private_key_file: The location of your private key file (RSA-SHA1 signature method only)
-     * - private_key_content: The content of your private key (RSA-SHA1 signature method only)
-     * - private_key_passphrase: The passphrase for your private key file (RSA-SHA1 signature method only)
+     * - private_key_file: The location of your private key file (RSA-SHA1
+     *   signature method only)
+     * - private_key_content: The content of your private key (RSA-SHA1
+     *   signature method only)
+     * - private_key_passphrase: The passphrase for your private key file
+     *   (RSA-SHA1 signature method only)
      * - token: Client token
      * - token_secret: Client secret token
      * - verifier: OAuth verifier.
      * - version: OAuth version. Defaults to '1.0'.
      * - realm: OAuth realm.
-     * - signature_method: Signature method. One of 'HMAC-SHA1', 'RSA-SHA1', or
-     *   'PLAINTEXT'. Defaults to 'HMAC-SHA1'.
+     * - signature_method: Signature method. One of 'HMAC-SHA1', 'RSA-SHA1',
+     *   'HMAC-SHA256', or 'PLAINTEXT'. Defaults to 'HMAC-SHA1'.
      *
      * @param array $config Configuration array.
      */
-    public function __construct($config)
+    public function __construct(array $config)
     {
         $this->config = [
-            'version'          => '1.0',
-            'request_method'   => self::REQUEST_METHOD_HEADER,
-            'consumer_key'     => 'anonymous',
-            'consumer_secret'  => 'anonymous',
+            'version' => '1.0',
+            'request_method' => self::REQUEST_METHOD_HEADER,
+            'consumer_key' => 'anonymous',
+            'consumer_secret' => 'anonymous',
             'signature_method' => self::SIGNATURE_METHOD_HMAC,
         ];
 
@@ -74,14 +80,14 @@ class Oauth1
     /**
      * Called when the middleware is handled.
      *
-     * @param callable $handler
-     *
      * @return \Closure
+     *
+     * @throws \InvalidArgumentException
+     * @throws \RuntimeException
      */
     public function __invoke(callable $handler)
     {
         return function ($request, array $options) use ($handler) {
-
             if (isset($options['auth']) && $options['auth'] == 'oauth') {
                 $request = $this->onBefore($request);
             }
@@ -90,12 +96,13 @@ class Oauth1
         };
     }
 
-    private function onBefore(RequestInterface $request)
+    /**
+     * @throws \InvalidArgumentException
+     * @throws \RuntimeException
+     */
+    private function onBefore(RequestInterface $request): RequestInterface
     {
-        $oauthparams = $this->getOauthParams(
-            $this->generateNonce($request),
-            $this->config
-        );
+        $oauthparams = self::getOauthParams($this->config);
 
         $oauthparams['oauth_signature'] = $this->getSignature($request, $oauthparams);
         uksort($oauthparams, 'strcmp');
@@ -126,11 +133,9 @@ class Oauth1
      * @param RequestInterface $request Request to generate a signature for
      * @param array            $params  Oauth parameters.
      *
-     * @return string
-     *
      * @throws \RuntimeException
      */
-    public function getSignature(RequestInterface $request, array $params)
+    public function getSignature(RequestInterface $request, array $params): string
     {
         // Remove oauth_signature if present
         // Ref: Spec: 9.1.1 ("The oauth_signature parameter MUST be excluded.")
@@ -148,7 +153,7 @@ class Oauth1
 
         $baseString = $this->createBaseString(
             $request,
-            $this->prepareParameters($params)
+            self::prepareParameters($params)
         );
 
         // Implements double-dispatch to sign requests
@@ -166,26 +171,11 @@ class Oauth1
                 $signature = $this->signUsingPlaintext($baseString);
                 break;
             default:
-                throw new \RuntimeException('Unknown signature method: ' . $this->config['signature_method']);
+                throw new \RuntimeException('Unknown signature method: '.$this->config['signature_method']);
                 break;
         }
 
         return base64_encode($signature);
-    }
-
-    /**
-     * Returns a Nonce Based on the unique id and URL.
-     *
-     * This will allow for multiple requests in parallel with the same exact
-     * timestamp to use separate nonce's.
-     *
-     * @param RequestInterface $request Request to generate a nonce for
-     *
-     * @return string
-     */
-    public function generateNonce(RequestInterface $request)
-    {
-        return sha1(uniqid('', true) . $request->getUri()->getHost() . $request->getUri()->getPath());
     }
 
     /**
@@ -198,28 +188,20 @@ class Oauth1
      * @param RequestInterface $request Request being signed
      * @param array            $params  Associative array of OAuth parameters
      *
-     * @return string Returns the base string
-     * @link http://oauth.net/core/1.0/#sig_base_example
+     * @see https://oauth.net/core/1.0/#sig_base_example
      */
-    protected function createBaseString(RequestInterface $request, array $params)
+    protected function createBaseString(RequestInterface $request, array $params): string
     {
         // Remove query params from URL. Ref: Spec: 9.1.2.
-        $url = $request->getUri()->withQuery('');
-        $query = http_build_query($params, '', '&', PHP_QUERY_RFC3986);
-
         return strtoupper($request->getMethod())
-            . '&' . rawurlencode($url)
-            . '&' . rawurlencode($query);
+            .'&'.rawurlencode((string) $request->getUri()->withQuery(''))
+            .'&'.rawurlencode(Query::build($params));
     }
 
     /**
-     * Convert booleans to strings, removed unset parameters, and sorts the array
-     *
-     * @param array $data Data array
-     *
-     * @return array
+     * @param array $data The data array
      */
-    private function prepareParameters($data)
+    private static function prepareParameters(array $data): array
     {
         // Parameters are sorted by name, using lexicographical byte value
         // ordering. Ref: Spec: 9.1.1 (1).
@@ -236,13 +218,10 @@ class Oauth1
 
     /**
      * @param string $algo Name of selected hashing algorithm (i.e. "md5", "sha256", "haval160,4", etc..)
-     * @param string $baseString
-     *
-     * @return string
      */
-    private function signUsingHmac($algo, $baseString)
+    private function signUsingHmac(string $algo, string $baseString): string
     {
-        $key = rawurlencode($this->config['consumer_secret']) . '&';
+        $key = rawurlencode($this->config['consumer_secret']).'&';
         if (isset($this->config['token_secret'])) {
             $key .= rawurlencode($this->config['token_secret']);
         }
@@ -251,15 +230,12 @@ class Oauth1
     }
 
     /**
-     * @param string $baseString
-     *
-     * @return string
+     * @throws RuntimeException
      */
-    private function signUsingRsaSha1($baseString)
+    private function signUsingRsaSha1(string $baseString): string
     {
         if (!function_exists('openssl_pkey_get_private')) {
-            throw new \RuntimeException('RSA-SHA1 signature method '
-                . 'requires the OpenSSL extension.');
+            throw new \RuntimeException('RSA-SHA1 signature method requires the OpenSSL extension.');
         }
 
         if (empty($this->config['private_key_content'])) {
@@ -281,11 +257,9 @@ class Oauth1
     }
 
     /**
-     * @param string $baseString
-     *
      * @return string
      */
-    private function signUsingPlaintext($baseString)
+    private function signUsingPlaintext(string $baseString)
     {
         return $baseString;
     }
@@ -294,51 +268,46 @@ class Oauth1
      * Builds the Authorization header for a request
      *
      * @param array $params Associative array of authorization parameters.
-     *
-     * @return array
      */
-    private function buildAuthorizationHeader(array $params)
+    private function buildAuthorizationHeader(array $params): array
     {
         foreach ($params as $key => $value) {
-            $params[$key] = $key . '="' . rawurlencode($value) . '"';
+            $params[$key] = $key.'="'.rawurlencode((string) $value).'"';
         }
 
         if (isset($this->config['realm'])) {
             array_unshift(
                 $params,
-                'realm="' . rawurlencode($this->config['realm']) . '"'
+                'realm="'.rawurlencode($this->config['realm']).'"'
             );
         }
 
-        return ['Authorization', 'OAuth ' . implode(', ', $params)];
+        return ['Authorization', 'OAuth '.implode(', ', $params)];
     }
 
     /**
      * Get the oauth parameters as named by the oauth spec
      *
-     * @param string     $nonce  Unique nonce
-     * @param array      $config Configuration options of the plugin.
-     *
-     * @return array
+     * @param array $config Configuration options of the plugin.
      */
-    private function getOauthParams($nonce, array $config)
+    private static function getOauthParams(array $config): array
     {
         $params = [
-            'oauth_consumer_key'     => $config['consumer_key'],
-            'oauth_nonce'            => $nonce,
+            'oauth_consumer_key' => $config['consumer_key'],
+            'oauth_nonce' => bin2hex(random_bytes(20)),
             'oauth_signature_method' => $config['signature_method'],
-            'oauth_timestamp'        => time(),
+            'oauth_timestamp' => time(),
         ];
 
         // Optional parameters should not be set if they have not been set in
         // the config as the parameter may be considered invalid by the Oauth
         // service.
         $optionalParams = [
-            'callback'  => 'oauth_callback',
-            'token'     => 'oauth_token',
-            'verifier'  => 'oauth_verifier',
-            'version'   => 'oauth_version',
-            'bodyhash'  => 'oauth_body_hash'
+            'callback' => 'oauth_callback',
+            'token' => 'oauth_token',
+            'verifier' => 'oauth_verifier',
+            'version' => 'oauth_version',
+            'bodyhash' => 'oauth_body_hash',
         ];
 
         foreach ($optionalParams as $optionName => $oauthName) {